requirements.txt
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.2
kiwisolver==1.4.4
matplotlib==3.6.0
nidaqmx==0.6.4
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
pypylon==1.8.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqtgraph==0.12.4
pyserial==3.5
python-dateutil==2.8.2
PyVISA==1.12.0
PyYAML==6.0
scipy==1.9.1
six==1.16.0
typing_extensions==4.3.0
wincertstore==0.2
